+++ src/base.js
@@ -96,6 +96,11 @@ class Base extends EventEmitter {
   get tokenSecret() {
     return this._secret;
   }
+
+  remove(...params) {
+    console.warn('The remove helper has been deprecated in favor of del. `remove` will no longer be available after the next major version bump (3.0).'); // eslint-disable-line no-console
+    this.del(...params);
+  }
 }
 
 module.exports = Base;